# api/api.yaml
openapi: 3.0.0
info:
  title: HackerNews API
  version: "1.0.0"
servers:
  - url: 'https://asw-hackernwes-g11.herokuapp.com'
    description: Heroku server
...
paths:
  /tweets:
    get:
      summary: Llista tots les publicacions
      tags: 
        - Publicacions
      parameters:
      - name: orderby
        required: false
        in: query
        description: Ordre en què es volen llistar les publicacions
        schema:
          type: string
      - name: user_id
        required: false
        in: query
        description: Id de l'usuari
        schema:
          type: integer
          minimum: 1  
      - name: voted
        required: false
        in: query
        description: Indica si es volen les contribucions que l'usuari ha votat
        schema:
          type: boolean
      responses:
        200:
          description: Retorna totes les publicacions
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      example: 1
                    author:
                      type: string
                      example: user1
                    content:
                      type: string
                      example: ""
                    created_at:
                      type: string
                      example: "2021-05-08T00:52:46.169Z"
                    url:
                      type: string
                      example: www.fib.upc.edu
                    title:
                      type: string
                      example: La FIB
                  required:
                    - id
                    - author
                    - content
                    - created_at
                    - url
                    - title
      security:
        - ApiKeyAuth: []
        
    post:
      summary: Creació d'una publicació
      tags: 
        - Publicacions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: La FIB
                url:
                  type: string
                  example: www.fib.upc.edu
                content:
                  type: string
                  example: Aquesta és la meva facultat
              required:
                - title
      responses:
        201:
          description: Publicació creada correctament
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tweet'
        401:
          description: La crida no conté una api_key vàlida
          content:
            application/json:
              schema:
                type: string
                example: Unauthorized, you don't have a valid api_key
        400:
          description: "Error - Aquesta URL ja existeix, el titol està buit o s'ha omplert el content i la url a la vegada"
          content:
            application/json:
              schema:
                type: string
                example: The URL already exists
      security:
        - ApiKeyAuth: []
...
  /comments:
    get:
      summary: Llista tots els comentaris amb les caracteristiques demanades
      tags: 
        - Comentaris
      parameters:
        - name: user_id
          in: query
          description: Id de l'usuari
          schema:
            type: integer
            minimum: 1  
        - name: voted
          in: query
          description: Indica si es volen els comentaris que l'usuari ha votat
          schema:
            type: boolean
      responses:
        200:
          description: Retorna tots els comentaris
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Comment'
        403:
          description: Error - L'usuari no pot veure els comentaris votats d'altres usuaris
      security:
        - ApiKeyAuth: []
    post:
      summary: Realitzar un comentari
      tags:
        - Comentaris
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  type: string
                  example: Molt interessant!
                contribution:
                  type: integer
                  example: 1
                  minimum: 1
                escomment:
                  type: boolean
                  example: false
                comment_id:
                  type: integer
                  example: 1
              required:
                - contribution
                - text
      responses:
        201:
          description: El comentari s'ha creat correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        400:
          description: El camp text no pot ser blanc
          content:
            application/json:
              schema:
                type: string
                example: Text property cannot be blank
        403:
          description: L'usuari esta comentant a la seva propia contribució
          content:
            application/json:
              schema:
                type: string
                example: Un usuari no pot fer comentaris de la seva propia contribució
        422:
          description: No es pot crear un comentari amb aquest contingut
          content:
            application/json:
              schema:
                type: string
                example: Unprocessable entity
      security:
        - ApiKeyAuth: []
...
components:
  schemas:
    Tweet:
      type: object
      properties:
        id:
          type: integer
          example: 1
        author:
          type: string
          example: user1
        content:
          type: string
          example: ""
        created_at:
          type: string
          example: "2021-05-08T00:52:46.169Z"
        url:
          type: string
          example: www.fib.upc.edu
        title:
          type: string
          example: La FIB
      required:
        - id
        - author
        - content
        - created_at
        - url
        - title
    Comment:
      type: object
      properties:
        comment_id:
          type: integer
          minimum: 1
          example: 1
        text:
          type: string
          example: Hola
        user:
          type: integer
          minimum: 1
          example: 1
          description: Id de l'usuari que ha fet el comentari
        contribution:
          type: integer
          minimum: 1
          example: 1
          description: id de la contribucio a la que fa referencia
        created_at:
          type: string
          example: "2021-05-05T17:18:47.855Z"
        points:
          type: integer
          minimum: 1
          example: 1
        escomment:
          type: boolean
        updated_at:
          type: string
          example: "2021-05-05T17:18:47.855Z"
      required:
        - comment_id
        - text
        - user
        - contribution
        - created_at
        - points
        - escomment
        - updated_at
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      name: X-API-KEY
      in: header